# storage/swagger/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Desafio Técnico - Backend PHP 2023",
        "description": "Desafio Técnico - Backend PHP 2023",
        "contact": {
            "email": "acarlos.os@hotmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://desafio.teste/api",
            "description": "local"
        }
    ],
    "paths": {
...
        "/task/{id}": {
            "get": {
                "tags": [
                    "Task"
                ],
                "summary": "Receive a task",
                "description": "Receive a task",
                "operationId": "task-show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "type": {
                                                    "properties": {
                                                        "id": {
                                                            "type": "integer",
                                                            "example": "1"
                                                        },
                                                        "type_id": {
                                                            "type": "integer",
                                                            "example": "1"
                                                        },
                                                        "user_id": {
                                                            "type": "integer",
                                                            "example": "1"
                                                        },
                                                        "title": {
                                                            "type": "string",
                                                            "example": "custom"
                                                        },
                                                        "description": {
                                                            "type": "string",
                                                            "example": "custom"
                                                        },
                                                        "start_date": {
                                                            "type": "string",
                                                            "format": "date",
                                                            "example": "custom"
                                                        },
                                                        "deadline": {
                                                            "type": "string",
                                                            "format": "date",
                                                            "example": "custom"
                                                        },
                                                        "finish_date": {
                                                            "type": "string",
                                                            "format": "date",
                                                            "example": "custom"
                                                        },
                                                        "status": {
                                                            "type": "string",
                                                            "example": "open"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "success": {
                                            "type": "boolean",
                                            "example": "true"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Comment updated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Wrong error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": "false"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Sorry, wrong error. Please try again"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Task"
                ],
                "summary": "Delete task",
                "description": "Delete task",
                "operationId": "task-delete",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The resource was deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": "true"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "The resource was deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Wrong error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": "false"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Sorry, wrong error. Please try again"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Laravel Sanctum token authentication",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}